requirements.txt
PyYAML==3.11
click==4.1
futures==3.0.3
kafka-python==0.9.4
kazoo==2.2.1
protobuf==2.6.1
psycopg2-managed-connection==1.0.0
psycopg2==2.6.1
pytest-capturelog==0.7
pytest==2.7.2
raven==5.5.0
setuptools==18.1
tabulate==0.7.5
